Python/Snowflake/requirements.txt
snowflake-connector-python==2.2.3